@@ -7,36 +7,35 @@ import _ from 'lodash';
 import { paramsCheck } from './utils';
 
 class HttpProvider {
-
   constructor(urlString) {
     this.url = url.parse(urlString);
     this.reqId = 0;
   }
 
   request(params) {
-
     // Make sure method is defined in params
     paramsCheck('request', params, ['method']);
 
-    let bodyJson = _.extend({
+    // Construct body of request options
+    const bodyJson = _.extend({
       id: this.reqId,
       jsonrpc: '1.0',
       method: '',
       params: [],
-    }, params)
+    }, params);
 
-    let reqOpts = {
+    // Construct options of request
+    const reqOpts = {
       method: 'POST',
       headers: {
         'Content-Type': 'text/plain',
       },
-      body: JSON.stringify(bodyJson)
-    }
+      body: JSON.stringify(bodyJson),
+    };
 
-    console.log(`${this.url.protocol}//${this.url.host}`, reqOpts);
-
+    // Add Basic Auth header if auth is defined in HttpProvider constructor
     if (this.url.auth) {
-      reqOpts.headers["Authorization"] = "Basic " + new Buffer(this.url.auth).toString("base64");
+      reqOpts.headers.Authorization = `Basic ${Buffer.from(this.url.auth).toString('base64')}`;
     }
 
     this.reqId += 1;
@@ -49,13 +48,12 @@ class HttpProvider {
 
   /**
    * Returns resolved Promise if Http response contains result; otherwise returns rejected upon error.
-   * 
+   *
    * @param  {object} response   JSON response from a HTTP request
    *
    * @return {object|undefined} Returns either the response, or throws an error
    */
   checkStatus(response) {
-
     // We can rely on checking error object so dont check HTTP status code here.
     if (response.error) {
       throw new Error(response.error.message);
