@@ -1,11 +1,10 @@
-import Qtum from './qtum.js';
-import HttpProvider from './httpprovider.js';
+/* Internal Import */
+import Qtum from './qtum';
+import HttpProvider from './httpprovider';
 import Contract from './contract';
 
 class Qweb3 {
-
   constructor(url) {
-
     const self = this;
 
     // TODO: url string validation
@@ -15,21 +14,27 @@ class Qweb3 {
     this.Contract = (address, abi) => new Contract(self, address, abi);
   }
 
+  /**
+   * Returns true if getinfo request returns result; otherwise false
+   * @return {Boolean}
+   */
   isConnected() {
     return this.provider.request({
-        method: 'getinfo'
-      })
-      .then((res) => {
-        return Promise.resolve(!!res);
-      }, (err) => {
-        return Promise.resolve(false);
-      });
+      method: 'getinfo',
+    })
+      .then(res => Promise.resolve(!!res), err => Promise.resolve(false));
   }
 
-  getTransaction(txid){
+  /**
+   * Get transaction details by txid
+   * @param  {string} txid transaction Id (64 digits hexString),
+   *                       e.g. dfafd59050fbe825d884b1e9279924f42bfa9506ca11e3d1910141054858f338
+   * @return {Promise}     Promise containing result object or Error
+   */
+  getTransaction(txid) {
     return this.provider.request({
       method: 'gettransaction',
-      params: [txid]
+      params: [txid],
     });
   }
 }
